NF: move all Compat tests in same package
This is the package of package implementation
