+++ AnkiDroid/src/test/java/com/ichi2/compat/CompatHasFilesTest.kt
@@ -14,13 +14,13 @@
  *  this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.ichi2.anki.compat
+package com.ichi2.compat
 
-import com.ichi2.compat.Compat
-import com.ichi2.compat.CompatV21
-import com.ichi2.compat.CompatV26
+import android.os.Build
 import com.ichi2.testutils.*
-import com.ichi2.testutils.AnkiAssert.assertDoesNotThrow
+import com.ichi2.testutils.createTransientDirectory
+import com.ichi2.testutils.withTempFile
+import org.hamcrest.CoreMatchers
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Test
@@ -29,14 +29,16 @@ import org.junit.runners.Parameterized
 import java.io.File
 import java.io.FileNotFoundException
 import java.io.IOException
+import java.nio.file.NotDirectoryException
 
-/** Tests for [Compat.deleteFile] */
+/** Tests for [Compat.hasFiles] */
 @RunWith(Parameterized::class)
-class CompatDeleteFileTest(
+class CompatHasFilesTest(
     val compat: Compat,
     /** Used in the "Test Results" Window */
     @Suppress("unused") private val unitTestDescription: String
 ) {
+
     companion object {
         @JvmStatic
         @Parameterized.Parameters(name = "{1}")
@@ -47,32 +49,34 @@ class CompatDeleteFileTest(
     }
 
     @Test
-    fun delete_file_which_exists() {
-        val file = createTransientFile()
-        assertDoesNotThrow { deleteFile(file) }
-        assertThat("file should no longer exist", file.exists(), equalTo(false))
+    fun has_files_with_file() {
+        val dir = createTransientDirectory().withTempFile("aa.txt")
+        assertThat("empty directory has no files", hasFiles(dir), equalTo(true))
     }
 
     @Test
-    fun delete_folder_which_exists() {
+    fun has_files_exists() {
         val dir = createTransientDirectory()
-        assertDoesNotThrow { deleteFile(dir) }
-        assertThat("directory should no longer exist", dir.exists(), equalTo(false))
+        assertThat("empty directory has no files", hasFiles(dir), equalTo(false))
     }
 
     @Test
-    fun delete_fails_if_exists_is_false() {
+    fun has_files_not_exists() {
         val dir = createTransientDirectory()
         dir.delete()
-        assertThrows<FileNotFoundException> { deleteFile(dir) }
+
+        assertThrows<FileNotFoundException> { hasFiles(dir) }
     }
 
     @Test
-    fun delete_fails_if_not_empty_directory() {
-        // Note: Exception is a DirectoryNotEmptyException in V26
-        val dir = createTransientDirectory().withTempFile("foo.txt")
-        assertThrowsSubclass<IOException> { deleteFile(dir) }
+    fun has_files_on_file() {
+        val file = createTransientFile("hello")
+
+        val exception = assertThrowsSubclass<IOException> { hasFiles(file) }
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            assertThat("Starting at API 26, this should be a NotDirectoryException", exception, CoreMatchers.instanceOf(NotDirectoryException::class.java))
+        }
     }
 
-    private fun deleteFile(file: File) = compat.deleteFile(file)
+    private fun hasFiles(dir: File) = compat.hasFiles(dir)
 }
